strange_data_types/requirements.txt
geojson==3.2.0
psycopg2-binary==2.9.11
